# adapter/src/main/resources/api.yaml
openapi: 3.0.3
info:
  title: Route Service API
  version: 1.0.0
  description: |
    Route planning: brouter proxy (semaphore-protected), draft persistence,
    elevation profile (Copernicus DEM). Plus admin endpoints to manage BRouter
    segments (.rd5) and profiles.
servers:
  - url: http://localhost:8085
    description: Local development server
paths:
...
  /admin/elevation/tiles:
    get:
      summary: List installed + available Copernicus DEM tiles (EU bbox by default)
      operationId: listDemTiles
      tags: [ElevationTiles]
      security: []
      parameters:
        - name: includeWorld
          in: query
          required: false
          description: If true, return all globe tiles. Default false = EU only.
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: Catalog of DEM tiles + installed flags
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DemTileResponse'
  /admin/elevation/tiles/download:
    post:
      summary: Start async download of a Copernicus DEM tile
      operationId: downloadDemTile
      tags: [ElevationTiles]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DownloadDemTileRequest'
      responses:
        '202':
          description: Task accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DemTileTaskResponse'
  /admin/elevation/tiles/{name}:
    delete:
      summary: Delete an installed DEM tile (.tif)
      operationId: deleteDemTile
      tags: [ElevationTiles]
      security: []
      parameters:
        - name: name
          in: path
          required: true
          description: Tile name without .tif (e.g. "N50E020")
          schema:
            type: string
      responses:
        '204':
          description: Deleted
...
  /admin/route/profiles/{name}:
    get:
      summary: Read profile content (plain text .brf)
      operationId: getProfile
      tags: [Profiles]
      parameters:
        - name: name
          in: path
          required: true
          description: Profile filename without .brf suffix
          schema:
            type: string
      responses:
        '200':
          description: Profile content
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProfileContent'
    put:
      summary: Save profile content (atomic write, no worker restart needed)
      operationId: saveProfile
      tags: [Profiles]
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SaveProfileRequest'
      responses:
        '204':
          description: Saved
    delete:
      summary: Delete a profile
      operationId: deleteProfile
      tags: [Profiles]
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Deleted

  # ===================== Planning (asystent tras) =====================
  /planning/session:
    get:
      summary: Get current planning session (intent + preferences + days)
      operationId: getPlanningSession
      tags: [Planning]
      responses:
        '200':
          description: Active session for authenticated user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlanningSessionResponse'
        '204':
          description: No session yet (user never entered the assistant)
        '401':
          description: Unauthorized
    delete:
      summary: Reset session ("Start over") — clears intent, preferences, days
      operationId: resetPlanningSession
      tags: [Planning]
      responses:
        '204':
          description: Session reset
        '401':
          description: Unauthorized
  /planning/session/intent:
    put:
      summary: Set or change intent (COVERAGE / AB / FREESTYLE) — clears preferences + days
      operationId: setPlanningIntent
      tags: [Planning]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SetPlanningIntentRequest'
      responses:
        '200':
          description: Updated session
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlanningSessionResponse'
        '400':
          description: Invalid intent
        '401':
          description: Unauthorized
  /planning/session/form:
    put:
      summary: Patch form preferences (non-null fields are merged in)
      operationId: updatePlanningForm
      tags: [Planning]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RoutePreferences'
      responses:
        '200':
          description: Updated session
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlanningSessionResponse'
        '401':
          description: Unauthorized
  /planning/calculate:
    post:
      summary: Submit async PLAN_ROUTE task — returns taskId for SSE/polling
      operationId: calculatePlan
      tags: [Planning]
      responses:
        '202':
          description: Task accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CalculatePlanResponse'
        '400':
          description: Form incomplete (missing fields for the chosen intent)
        '401':
          description: Unauthorized
        '409':
          description: Plan task already running / queue full
  /planning/cancel:
    post:
      summary: Cancel active PLAN_ROUTE task (no-op if none)
      operationId: cancelPlan
      tags: [Planning]
      responses:
        '204':
          description: Cancellation signal sent
        '401':
          description: Unauthorized
  /planning/task/{taskId}:
    get:
      summary: Get current status of a planning task (polling endpoint for frontend progress)
      operationId: getPlanningTask
      tags: [Planning]
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Task snapshot (status + phase + progress)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlanTask'
        '401':
          description: Unauthorized
        '404':
          description: Task not found
  /planning/session/days/{dayNumber}:
    put:
      summary: Save edited geometry/waypoints for one day (re-routes via BRouter + DEM)
      operationId: updatePlanningDay
      tags: [Planning]
      parameters:
        - name: dayNumber
          in: path
          required: true
          schema:
            type: integer
            minimum: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePlanningDayRequest'
      responses:
        '200':
          description: Day re-routed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlanningSessionDay'
        '401':
          description: Unauthorized
        '404':
          description: Day not found
  /planning/session/save-as-expedition:
    post:
      summary: Convert all days to route_draft entries with shared group_id
      operationId: savePlanAsExpedition
      tags: [Planning]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SaveAsExpeditionRequest'
      responses:
        '200':
          description: Drafts created (one per day)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SaveAsExpeditionResponse'
        '401':
          description: Unauthorized
components:
  schemas:
...
    DemTileResponse:
      type: object
      required: [name, lonStart, latStart, installed]
      properties:
        name:
          type: string
          description: e.g. "N50E020" (no .tif suffix)
          example: N50E020
        lonStart:
          type: integer
        latStart:
          type: integer
        remoteSizeBytes:
          type: integer
          format: int64
          nullable: true
        installed:
          type: boolean
        installedSizeBytes:
          type: integer
          format: int64
          nullable: true
        installedAt:
          type: string
          format: date-time
          nullable: true
        outdated:
          type: boolean
          nullable: true
    DownloadDemTileRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          example: N50E020
    DemTileTaskResponse:
      type: object
      required: [taskId, tileName, status]
      properties:
        taskId:
          type: string
        tileName:
          type: string
        status:
          type: string
          enum: [QUEUED, RUNNING, COMPLETED, FAILED, CANCELLED]
...
    ProfileContent:
      type: object
      required: [name, content]
      properties:
        name:
          type: string
        content:
          type: string
          description: Full text content of the .brf file
        sizeBytes:
          type: integer
          format: int64
        modifiedAt:
          type: string
          format: date-time
          nullable: true
    SaveProfileRequest:
      type: object
      required: [content]
      properties:
        content:
          type: string

    # ===================== Planning DTOs =====================
    PlanningIntent:
      type: string
      enum: [COVERAGE, AB, FREESTYLE]
    PlanningRouteStyle:
      type: string
      enum: [KRAJOWKI, SCIEZKI, GRAVEL, WZDLUZ_RZEK, PRZEZ_MIASTA]
    PlanningTempo:
      type: string
      enum: [TURYSTYCZNIE, SZYBKO]
    Waypoint:
      type: object
      required: [lng, lat]
      properties:
        lng:
          type: number
          format: double
        lat:
          type: number
          format: double
        name:
          type: string
          nullable: true
    RoutePreferences:
      type: object
      description: All fields nullable — PATCH semantics. Lists replace in full when provided.
      properties:
        countryIds:
          type: array
          nullable: true
          items:
            type: integer
        levelIds:
          type: array
          nullable: true
          items:
            type: integer
        specialGroupIds:
          type: array
          nullable: true
          items:
            type: integer
        start:
          $ref: '#/components/schemas/Waypoint'
        end:
          $ref: '#/components/schemas/Waypoint'
        via:
          type: array
          nullable: true
          items:
            $ref: '#/components/schemas/Waypoint'
        loop:
          type: boolean
          nullable: true
        days:
          type: integer
          nullable: true
          minimum: 1
          maximum: 20000
        kmPerDay:
          type: integer
          nullable: true
          minimum: 1
          maximum: 1000
        elevationPerDayM:
          type: integer
          nullable: true
          minimum: 0
          maximum: 10000
        profile:
          type: string
          nullable: true
          description: Direct BRouter profile name (fastbike/trekking/safety/fastbike-lowtraffic)
        style:
          $ref: '#/components/schemas/PlanningRouteStyle'
        tempo:
          $ref: '#/components/schemas/PlanningTempo'
    PlanningSessionDay:
      type: object
      required: [id, sessionId, dayNumber, geometryEncoded, waypoints, profile, editedAt]
      properties:
        id:
          type: string
          format: uuid
        sessionId:
          type: string
          format: uuid
        dayNumber:
          type: integer
        geometryEncoded:
          type: string
          description: Polyline3DCodec encoded (lng,lat,z)
        waypoints:
          type: array
          items:
            $ref: '#/components/schemas/Waypoint'
        distanceKm:
          type: number
          format: double
          nullable: true
        elevationGain:
          type: integer
          nullable: true
        elevationLoss:
          type: integer
          nullable: true
        profile:
          type: string
        editedAt:
          type: string
          format: date-time
    PlanningSessionResponse:
      type: object
      required: [id, userId, preferences, days, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        intent:
          $ref: '#/components/schemas/PlanningIntent'
        preferences:
          $ref: '#/components/schemas/RoutePreferences'
        days:
          type: array
          items:
            $ref: '#/components/schemas/PlanningSessionDay'
        lastTaskId:
          type: string
          format: uuid
          nullable: true
        summary:
          $ref: '#/components/schemas/PlanningSummary'
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    SetPlanningIntentRequest:
      type: object
      required: [intent]
      properties:
        intent:
          $ref: '#/components/schemas/PlanningIntent'
    CalculatePlanResponse:
      type: object
      required: [taskId]
      properties:
        taskId:
          type: string
          format: uuid
          description: UUID of submitted PLAN_ROUTE task (poll via GET /planning/task/{taskId})
    UpdatePlanningDayRequest:
      type: object
      required: [waypoints]
      properties:
        waypoints:
          type: array
          minItems: 2
          items:
            $ref: '#/components/schemas/Waypoint'
    SaveAsExpeditionRequest:
      type: object
      required: [groupName]
      properties:
        groupName:
          type: string
          minLength: 1
          maxLength: 200
    SaveAsExpeditionResponse:
      type: object
      required: [draftIds, groupId]
      properties:
        groupId:
          type: string
          format: uuid
        draftIds:
          type: array
          items:
            type: string
            format: uuid
    PlanningSummary:
      type: object
      description: Podsumowanie policzonej wyprawy. Null gdy sesja jeszcze nie była policzona.
      required: [totalDistanceKm, totalElevationGain, budgetKm, verdict, surplusKm, poolSize, initialPoolSize, reconcileIters, reconcileTrims, reconcileGrows]
      properties:
        totalDistanceKm:    { type: number, format: double }
        totalElevationGain: { type: integer }
        budgetKm:           { type: integer, description: "days × kmPerDay" }
        verdict:
          type: string
          enum: [OK, SURPLUS, DEFICIT, BUDGET_IMPOSSIBLE]
          description: "BUDGET_IMPOSSIBLE = sam baseline (start→via→meta) przekracza budżet — user musi poprawić parametry."
        surplusKm:          { type: integer }
        poolSize:           { type: integer, description: "Liczba obszarów po reconcile (finalna pula)" }
        initialPoolSize:    { type: integer, description: "Liczba obszarów PRZED reconcile (cap MAX_TSP_AREAS)" }
        reconcileIters:     { type: integer }
        reconcileTrims:     { type: integer }
        reconcileGrows:     { type: integer }
        baselineKm:
          type: number
          format: double
          nullable: true
          description: "Baseline BRouter (start→via→meta bez gmin). Dolna granica trasy."
        roadAnchors:
          type: number
          format: double
          nullable: true
          description: "Współczynnik road/straight dla anchors-only (probe baseline). Diagnostyka."
        roadAreas:
          type: number
          format: double
          nullable: true
          description: "Współczynnik road/straight dla gęstej puli obszarów (density probe). Diagnostyka."
        climbWarning:
          type: boolean
          description: "true gdy totalElevationGain > refClimbTotal × 1.10 — user-facing warning, nie blokuje verdict."
    PlanTaskStatus:
      type: string
      enum: [RUNNING, COMPLETED, FAILED, CANCELLED]
    PlanTask:
      type: object
      required: [id, sessionId, userId, status, startedAt]
      properties:
        id:
          type: string
          format: uuid
        sessionId:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        status:
          $ref: '#/components/schemas/PlanTaskStatus'
        phase:
          type: string
          nullable: true
          description: Current phase name (validating / fetching-areas / ordering-areas / routing-brouter / sampling-elevation / splitting-days / computing-day-N / saving / done)
        progressCurrent:
          type: integer
          nullable: true
        progressTotal:
          type: integer
          nullable: true
        error:
          type: string
          nullable: true
        startedAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
          nullable: true